Add logger class to common lib to enable logging info with protected std::out and added timestamp. Fix exception handling
